requirements-build.txt
# Maintainer-only: Windows folder build (PyInstaller). Not required to run from source.
-r requirements.txt
pyinstaller==6.11.1
Pillow>=10.0.0
